functions/collective/wiki_css_backup/requirements.txt
praw==7.0.0
prawcore==1.4.0
python-gitlab==2.0.0
